Revert back by mistake the maxWidth

+++ src/claviska/SimpleImage.php
@@ -1376,7 +1376,7 @@ class SimpleImage
     /**
      * Receives a text and breaks into LINES.
      */
-    private function textSeparateLines(string $text, string $fontFile, float $fontSize, float $maxWidth): array
+    private function textSeparateLines(string $text, string $fontFile, float $fontSize, int $maxWidth): array
     {
         $lines = [];
         $words = self::textSeparateWords($text);
